# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sovren API
  description: |
    **Sovren API**: NOSTR-native decentralized creator monetization platform

    ## Overview
    The Sovren API provides comprehensive endpoints for content management, user profiles,
    and payment processing via Bitcoin Lightning Network. Built on the NOSTR protocol
    for true decentralization and censorship resistance.

    ## Base URLs
    - **Development**: `http://localhost:3001/api/v1`
    - **Production**: `https://api.sovren.app/v1`

    ## Authentication
    All authenticated endpoints require a JWT bearer token obtained via the authentication flow.
    The token must include the user's NOSTR public key for protocol compliance.

    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limiting
    API requests are rate-limited per endpoint category:
    - Content operations: 100 requests/hour
    - User operations: 200 requests/hour
    - Payment operations: 50 requests/hour

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Total requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Time when limit resets (Unix timestamp)

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes:
    - `400`: Bad Request - Validation error
    - `401`: Unauthorized - Authentication required
    - `403`: Forbidden - Insufficient permissions
    - `404`: Not Found - Resource doesn't exist
...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /content/publish:
    post:
      tags: [Content]
      summary: Publish new content
      description: |
        Publishes new content to the platform and optionally to NOSTR relays.
        Supports multiple content types including articles, videos, audio, images, and raw NOSTR events.
      operationId: publishContent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishContentRequest'
            examples:
              article:
                summary: Publish an article
                value:
                  title: "Getting Started with Bitcoin Lightning"
                  content: "In this comprehensive guide, we'll explore..."
                  contentType: "article"
                  tags: ["bitcoin", "lightning", "tutorial"]
                  publishToNostr: true
                  monetization:
                    priceInSats: 1000
                    allowTipping: true
              video:
                summary: Publish a video
                value:
                  title: "NOSTR Protocol Explained"
                  content: "https://cdn.sovren.app/videos/nostr-explained.mp4"
                  contentType: "video"
                  tags: ["nostr", "protocol", "education"]
                  monetization:
                    priceInSats: 5000
                    allowTipping: true
                    subscriptionTier: "premium"
      responses:
        '201':
          description: Content published successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishContentResponse'
              examples:
                success:
                  value:
                    success: true
                    data:
                      contentId: "550e8400-e29b-41d4-a716-446655440000"
                      status: "published"
                      nostrEventId: "7f9e8d7c6b5a4f3e2d1c0b9a8f7e6d5c4b3a2f1e0d9c8b7a6f5e4d3c2b1a0"
                      publishedAt: "2025-10-27T10:30:00Z"
                      url: "https://sovren.app/content/550e8400-e29b-41d4-a716-446655440000"
                      relayPublishResults:
                        - relayUrl: "wss://relay.damus.io"
                          success: true
                        - relayUrl: "wss://relay.nostr.band"
                          success: true
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /content/moderate:
    post:
      tags: [Content]
      summary: Moderate content
      description: |
        Performs moderation actions on published content. Available actions include approve, reject,
        flag for review, or mark for manual review. Supports both manual and AI-assisted moderation.
      operationId: moderateContent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerateContentRequest'
            examples:
              approve:
                summary: Approve content
                value:
                  contentId: "550e8400-e29b-41d4-a716-446655440000"
                  action: "approve"
                  moderatorId: "mod-123"
                  notes: "Content meets community guidelines"
              flag:
                summary: Flag content for review
                value:
                  contentId: "550e8400-e29b-41d4-a716-446655440001"
                  action: "flag"
                  reason: "Potential copyright violation"
                  moderatorId: "mod-456"
      responses:
        '200':
          description: Content moderated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerateContentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /content/search:
    get:
      tags: [Content]
      summary: Search content
      description: |
        Performs full-text search across published content with advanced filtering and sorting options.
        Supports filtering by content type, tags, author, date range, and price range.
      operationId: searchContent
      parameters:
        - name: query
          in: query
          required: true
          description: Search query string
          schema:
            type: string
            minLength: 1
            maxLength: 500
          example: "bitcoin lightning network"
        - name: contentType
          in: query
          description: Filter by content type (comma-separated)
          schema:
            type: string
          example: "article,video"
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
          example: "bitcoin,nostr"
        - name: authorId
          in: query
          description: Filter by author user ID
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          description: Filter by start date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          description: Filter by end date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: minPrice
          in: query
          description: Minimum price in sats
          schema:
            type: integer
            minimum: 0
        - name: maxPrice
          in: query
          description: Maximum price in sats
          schema:
            type: integer
            minimum: 0
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [relevance, date, popularity, price]
            default: relevance
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchContentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /content/recommendations:
    get:
      tags: [Content]
      summary: Get content recommendations
      description: |
        Returns AI-powered personalized content recommendations based on user interests,
        viewing history, and engagement patterns.
      operationId: getRecommendations
      parameters:
        - name: userId
          in: query
          required: true
          description: User ID for personalization
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Number of recommendations to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: diversity
          in: query
          description: Recommendation diversity (0-1 scale)
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
            default: 0.5
      responses:
        '200':
          description: Content recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRecommendationsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /content/analytics/{id}:
    get:
      tags: [Content]
      summary: Get content analytics
      description: |
        Retrieves comprehensive analytics for a specific piece of content including views,
        engagement, revenue, traffic sources, and performance metrics.
      operationId: getContentAnalytics
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Content ID
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          description: Start date for analytics (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date for analytics (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: metrics
          in: query
          description: Specific metrics to retrieve (comma-separated)
          schema:
            type: string
          example: "views,engagement,revenue"
      responses:
        '200':
          description: Content analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentAnalyticsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /content/versions/{id}:
    get:
      tags: [Content]
      summary: Get version history
      description: |
        Retrieves complete version history for a piece of content including all changes,
        timestamps, and authors.
      operationId: getVersionHistory
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Content ID
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Number of versions to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of versions to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Version history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVersionHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /content/versions/{id}/revert:
    post:
      tags: [Content]
      summary: Revert to previous version
      description: |
        Reverts content to a previous version. Creates a new version entry while preserving
        the complete version history.
      operationId: revertContentVersion
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Content ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevertContentVersionRequest'
            examples:
              revert:
                value:
                  contentId: "550e8400-e29b-41d4-a716-446655440000"
                  targetVersionId: "ver-789"
                  reason: "Reverting to fix formatting issues introduced in latest version"
      responses:
        '200':
          description: Content reverted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevertContentVersionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

  # ============================================================================
  # USER API ENDPOINTS (8 endpoints)
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the authentication endpoint.

        Token must include:
        - User ID
        - NOSTR public key
        - User role (supporter, creator, admin)
        - Expiration timestamp

        Example:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
  headers:
    X-RateLimit-Limit:
      description: Total number of requests allowed in the current window
      schema:
        type: integer
      example: 100
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
      example: 87
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
      example: 1698415200
  schemas:
...
    PublishContentRequest:
      type: object
      required:
        - title
        - content
        - contentType
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Content title
          example: "Getting Started with Bitcoin Lightning"
        content:
          type: string
          minLength: 1
          maxLength: 1000000
          description: Content body (text, markdown, or URL to media)
          example: "In this comprehensive guide..."
        contentType:
          type: string
          enum: [article, video, audio, image, nostr-event]
          description: Type of content being published
          example: "article"
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 50
          maxItems: 20
          description: Content tags for categorization and search
          example: ["bitcoin", "lightning", "tutorial"]
        metadata:
          type: object
          additionalProperties: true
          description: Additional content metadata
        publishToNostr:
          type: boolean
          default: true
          description: Whether to publish to NOSTR relays
        relayUrls:
          type: array
          items:
            type: string
            format: uri
          description: Specific NOSTR relay URLs (uses defaults if omitted)
          example: ["wss://relay.damus.io", "wss://relay.nostr.band"]
        monetization:
          type: object
          properties:
            priceInSats:
              type: integer
              minimum: 0
              description: Content price in satoshis (0 for free)
              example: 1000
            allowTipping:
              type: boolean
              default: true
              description: Allow tips on free content
            subscriptionTier:
              type: string
              description: Required subscription tier (if applicable)
              example: "premium"
    PublishContentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            contentId:
              type: string
              format: uuid
              description: Unique content identifier
              example: "550e8400-e29b-41d4-a716-446655440000"
            status:
              type: string
              enum: [published, pending, failed]
              description: Publication status
              example: "published"
            nostrEventId:
              type: string
              description: NOSTR event ID (if published to NOSTR)
              example: "7f9e8d7c6b5a4f3e2d1c0b9a8f7e6d5c4b3a2f1e0d9c8b7a6f5e4d3c2b1a0"
            relayPublishResults:
              type: array
              items:
                type: object
                properties:
                  relayUrl:
                    type: string
                    format: uri
                  success:
                    type: boolean
                  error:
                    type: string
              description: Results from each relay
            publishedAt:
              type: string
              format: date-time
              example: "2025-10-27T10:30:00Z"
            url:
              type: string
              format: uri
              description: Public URL for the content
              example: "https://sovren.app/content/550e8400-e29b-41d4-a716-446655440000"
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
  responses:
    ValidationError:
      description: Validation error - Request contains invalid data
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                success: false
                error:
                  code: "E001"
                  message: "Validation failed"
                  validationErrors:
                    - field: "title"
                      message: "Title is required"
                    - field: "amount"
                      message: "Amount must be greater than 0"
                metadata:
                  requestId: "req-123"
                  timestamp: "2025-10-27T10:30:00Z"
                  processingTime: 5.2
    Unauthorized:
      description: Unauthorized - Authentication required or token invalid
      headers:
        WWW-Authenticate:
          description: Authentication scheme
          schema:
            type: string
          example: Bearer realm="Sovren API"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_token:
              value:
                success: false
                error:
                  code: "E401"
                  message: "Authentication required"
                  details:
                    reason: "No bearer token provided"
                metadata:
                  requestId: "req-456"
                  timestamp: "2025-10-27T10:30:00Z"
            invalid_token:
              value:
                success: false
                error:
                  code: "E402"
                  message: "Invalid or expired token"
                  details:
                    reason: "Token signature verification failed"
...
    RateLimitExceeded:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
          example: 3600
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              value:
                success: false
                error:
                  code: "E429"
                  message: "Rate limit exceeded"
                  details:
                    limit: 100
                    window: "1 hour"
                    retryAfter: 3600
...